Modificado el precio a BigDecimal y agregado tests de integración

+++ zrQueryPrice-ms/src/main/java/com/croyan/queryprice/bean/ProductPriceBean.java
@@ -1,8 +1,9 @@
 package com.croyan.queryprice.bean;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record ProductPriceBean
         (Long brandId, Long productId, Long priceId,
-         LocalDateTime startDate, LocalDateTime endDate, double currentPrice) {
+         LocalDateTime startDate, LocalDateTime endDate, BigDecimal currentPrice) {
 }
